ARC-10: split feature.py into a subpackage (collection / geometry / crs)
The former single 955-line feature.py has been split into four modules
so no file is doing too many things at once:

  * pyramids/feature/collection.py — the FeatureCollection class
    (GeoDataFrame subclass, instance methods, pyramids-specific
    properties).
  * pyramids/feature/geometry.py — shape factories and coordinate
    extractors: create_polygon, create_point, get_xy_coords,
    get_point_coords, get_line_coords, get_poly_coords, explode_gdf,
    multi_geom_handler, geometry_collection_coords, get_coords.
  * pyramids/feature/crs.py — CRS/EPSG/reprojection helpers:
    create_sr_from_proj, get_epsg_from_prj, reproject_points (ARC-2
    Transformer-based), reproject_points_osr.
  * pyramids/feature/_ogr.py — private OGR bridge (unchanged).

The FeatureCollection class keeps thin static-method delegates for
every moved helper (create_polygon, create_point, get_epsg_from_prj,
reproject_points, reproject_points2, _create_sr_from_proj,
_get_xy_coords, _get_point_coords, _get_line_coords, _get_poly_coords,
_explode_gdf, _multi_geom_handler, _geometry_collection, _get_coords)
so existing callers that wrote 'FeatureCollection.create_polygon(...)'
continue to work without change.

The public import path is unchanged:

    from pyramids.feature import FeatureCollection

pyramids/feature/__init__.py now imports from collection.py.

Full non-plot suite: 2105 passed.
